Major WIP - Refactor logic to be single responsibility files, use async/await
